% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{count_upm}
\alias{count_upm}
\title{Count Methylation States}
\usage{
count_upm(data)
}
\arguments{
\item{data}{A numeric vector with methylation values: \code{0} (unmethylated), \code{0.5} (partially-methylated),
and \code{1} (methylated).}
}
\value{
An integer vector of length 3 containing counts of unmethylated, partially-methylated,
and methylated sites, respectively.
}
\description{
This internal function counts the number of sites with unmethylated, partially-methylated,
and methylated states in a given vector.
}
